tcti: Move 'commonchecks' module into 'tcti' module.
The two function in the 'commonchecks' module never really warranted
a separate module. As we refactor the TCTIs though we'll need a common
place to put utility functions in. We already have a 'tcti.h' with
utility macros and the contxt structure so it seems reasonable to me
to just combine the two into a single 'tcti' module. Any other convenience
functions and stuff should go here.
