+++ tcti/tcti.c
