+++ tcti/tcti.h
@@ -114,5 +114,22 @@ tcti_context_intel_cast (TSS2_TCTI_CONTEXT *ctx)
 {
     return (TSS2_TCTI_CONTEXT_INTEL*)ctx;
 }
+/*
+ * This function performs common checks on the context structure and the
+ * buffer passed into TCTI 'transmit' functions.
+ */
+TSS2_RC CommonSendChecks(
+    TSS2_TCTI_CONTEXT *tctiContext,
+    uint8_t           *command_buffer
+    );
+/*
+ * This function performs common checks on the context structure, buffer and
+ * size parameter passed to the TCTI 'receive' functions.
+ */
+TSS2_RC CommonReceiveChecks(
+    TSS2_TCTI_CONTEXT *tctiContext,
+    size_t            *response_size,
+    unsigned char     *response_buffer
+    );
 
 #endif
